// src/app/core/models/producto.interface.ts
export interface CreateProductoRequest {
  nombre: string;
  slug?: string;
  descripcion?: string;
  precio: number;
  precio_oferta?: number;
  stock: number;
  sku?: string;
  codigo_barras?: string;
  imagen_principal?: File;
  destacado?: boolean;
  activo?: boolean;
  categoria_id: number;
  marca?: string;
  modelo?: string;
  garantia?: string;
  meta_title?: string;
  meta_description?: string;
  idioma?: string;
  moneda?: string;
  atributos_extra?: Record<string, any>;
}